requirements.txt
humanize
PogoOCR==0.3.1
python-dateutil
Red-DiscordBot[postgres]==3.3.10
